Create daily_commit_20210416_python_numbers
Python Data Structure Quiz: Numbers
a)Write an equation that uses multiplication, division, an exponent, addition, and subtraction that is equal to 100.25.
b)Answer these 3 questions without typing code, then type code to check your answer. What is the value of the expression 4 * (6 + 5)? What is the value of the expression 4 * 6 + 5? What is the value of the expression 4 + 6 * 5?
c)What is the * type * of the result of the expression 3 + 1.5 + 4? 
d)What would you use to find a number's square root, as well as its square?