Create daily_commit_20210420_python_statements.py
Python Statement assessment test:
a)Use for, .split(), and if to create a Statement that will print out words that start with 's':
st = 'Print only the words that start with s in this sentence'
b)create this ['Print only the word', ' that ', 'tart with ', ' in thi', ' ', 'entence'] from the string in (a)
c)Use range() to print all the even numbers from 0 to 10.
d)Use a List Comprehension to create a list of all numbers between 1 and 50 that are divisible by 3.
e)Go through the string below and if the length of a word is even print "even!"
st3 = 'Print every word in this sentence that has an even number of letters'
f)Write a program that prints the integers from 1 to 100. But for multiples of three print "Fizz" instead of the number, and for the multiples of five print "Buzz". For numbers which are multiples of both three and five print "FizzBuzz".
g)Use List Comprehension to create a list of the first letters of every word in the string below:
st4 = 'Create a list of the first letters of every word in this string'