Create daily_commit_20210424_python_functions2.py
Python Functions Quiz 2:
a) SUMMER OF '69:
Return the sum of the numbers in the array,  except ignore sections of numbers starting with a 6 and extending to the next 9 (every 6 will be followed by at least one 9). Return 0 for no numbers. Example
 summer_69([1, 3, 5]) --> 9
 summer_69([4, 5, 6, 7, 8, 9]) --> 9
 summer_69([2, 1, 6, 9, 11]) --> 14

b) SPY GAME:
Write a function that takes in a list of integers and returns True if it contains 007 in order.Example
spy_game([1,2,4,0,0,7,5]) --> True
spy_game([1,0,2,4,0,5,7]) --> True
spy_game([1,7,2,0,4,5,0]) --> False

c) COUNT PRIMES:
Write a function that returns the number of prime numbers that exist up to and including a given number. By convention, 0 and 1 are not prime. Example
count_primes(100) --> 25

d) PRINT BIG:
Write a function that takes in a single letter, and returns a 5x5 representation of that letter. Example if you call : print_big('a')

output should be:
   *
  * *
*****
*     *
*     *
HINT: Consider making a dictionary of possible patterns, and mapping the alphabet to specific 5-line combinations of patterns. For purposes of this exercise, it's ok if your dictionary stops at "E".