daily_commit_20210624 reduce function
Given a list of rational numbers, find their product using the reduce function().
Input Format:
First line contains , the number of rational numbers.
The  of next  lines contain two integers each, the numerator(  ) and denominator(  ) of the  rational number in the list.

Constraints:
number of rational numbers is in the range 1 to 100
numerators and denominators are in the range 1 to 10E9

Output Format:
Print only one line containing the numerator and denominator of the product of the numbers in the list in its simplest form, i.e. numerator and denominator have no common divisor other than .