Create daily_commit_20210729_wordOrder.py
Input Format:
The first line contains the integer,n .
The next n lines each contain a word.

Output Format:
Output  lines;
On the first line, output the number of distinct words from the input.
On the second line, output the number of occurrences for each distinct word according to their appearance in the input.